backend/requirements.txt
Flask
flask-cors
requests
psycopg2
python-dotenv
openai>=1.0.0
gunicorn


